requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.0.1
Pillow==10.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
flask-bootstrap==3.3.7.1
openpyxl==3.1.2
PyMySQL==1.1.0